+++ src/widgets/Ecommerce.php
@@ -42,7 +42,7 @@ class Ecommerce extends \craft\base\Widget
     /**
      * @inheritdoc
      */
-    public static function iconPath()
+    public static function icon()
     {
         return Craft::getAlias('@dukt/analytics/icons/report.svg');
     }