+++ src/widgets/Realtime.php
@@ -54,7 +54,7 @@ class Realtime extends \craft\base\Widget
     /**
      * @inheritdoc
      */
-    public static function iconPath()
+    public static function icon()
     {
         return Craft::getAlias('@dukt/analytics/icons/realtime-report.svg');
     }